requirements.txt
copykitten==2.0.0
piper-tts==1.3.0
playsound3==3.3.1
customtkinter==5.2.2